# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "A-ROSA-JE API",
        "description": "Cool API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/plants": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Get all plants",
                "operationId": "26cdaf75d14d614ecc7fa69af4dbc493",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Plants"
                ],
                "summary": "Create a plant",
                "operationId": "be9d330153ac9ffc6e331b8aa2c8b697",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "specie_name": {
                                        "type": "string",
                                        "example": "My plant"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "My location"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "My status"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "My description"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/v1/plants/{id}": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Get an plants by ID",
                "operationId": "9922e0e38fc73eeb88019d8a41751b80",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Plants"
                ],
                "summary": "Update a plant by ID",
                "operationId": "9facc20f5f512e2c48f8d512ddd8fbf1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "specie_name": {
                                        "type": "string",
                                        "example": "My plant"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "My location"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "My status"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "My description"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Plants"
                ],
                "summary": "Delete a plant by ID",
                "operationId": "5949d336e2d34d7efde419117440e135",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
...
}